refactor: refactor configuration into one class per algo
The configuration is global and contains all configuration
parameters for all algorithms. It makes difficult to know the
values of the parameters used, in particular when algorithms have
parameters with identical names. It also creates the need to
specify an algorithm value in addition to a config, which must
match.

This commit refactors the configuration so each algorithm has its
own configuration object. All tests and cli are refactored
accordingly.

In addition, the FACT algorithm was completely removed.