feature: add large histogram generation for diffusion
This commit adds a new function in tg.algorithms.diffusion which
allows the generation of histograms with a very large number
of seeds. This overlaps with the purpose of Diffusion.histogram
which will eventually be deleted.

Some changes to the random number generation and seeds generation
needed for the histogram are also performed.