bugfix: points are not correctly detected in image
Some voxels space points, in particular seeds, are not always
correctly detected as in or out of the image. This is caused by
incorrect assumption about voxel centers.

This commit adopts the convention that voxel indices correspond
to the center of voxels, with valid points going from -0.5 to
N - 0.5 in voxel space.