feature: add histogram for diffusion
Current histograms are produced by generating streamlines and
then computing their distribution outside of the kernel. This is
inefficient and limits the number of seeds that can be used.

This commit allows very large histograms to be produced for the
diffusion algorithm by generating seeds directly in the kernel and
recording their distribution in realtime.